+++ fsk_demod.h
@@ -4,7 +4,7 @@
 int get_rtty_ch(int state);
 void setup_rx(void);
 void toggle_reverse(bool *rev);
-double get_waterfall(int bucket);
-void setup_spectrum_filters(int buckets);
+double get_waterfall(size_t bucket);
+void setup_spectrum_filters(size_t buckets);
 
 #endif
